refactor(product list): Uniform usage of image model component
